scramble: convert the abundance once, at the handle call
get_abundance() returns int64_t while OutputAnnotations::abundance is
uint64_t, so the fasta and fastq writers each carried their own
static_cast. Convert at the point of use instead: the cast now sits
next to the call it converts, which is where the signed-to-unsigned
question actually belongs, and the two call sites pass the value plain.

Value-preserving: get_abundance() returns the ;size= annotation when it
is positive, and 1 otherwise, so the operand is never negative. Same
form as cut.cpp and fastq_join.cpp already use.

No functional change.

+++ src/commands/scramble.cpp
@@ -435,7 +435,7 @@ auto scramble(struct Parameters const & parameters) -> void
         /* header */
 
         auto const header = input_handle->header_view();
-        auto const abundance = input_handle->get_abundance();
+        auto const abundance = static_cast<uint64_t>(input_handle->get_abundance());
 
 
         /* sequence */
@@ -473,7 +473,7 @@ auto scramble(struct Parameters const & parameters) -> void
                                 nullptr,
                                 make_view(seq_buffer).first(length),
                                 header,
-                                OutputAnnotations{static_cast<uint64_t>(abundance), count},
+                                OutputAnnotations{abundance, count},
                                 parameters);
           }
 
@@ -483,7 +483,7 @@ auto scramble(struct Parameters const & parameters) -> void
                                 make_view(seq_buffer).first(length),
                                 header,
                                 input_handle->quality_view(),
-                                OutputAnnotations{static_cast<uint64_t>(abundance), count},
+                                OutputAnnotations{abundance, count},
                                 parameters);
           }
